Move spam detection to new Module

+++ DirtBag/DirtBag/Modules/AnalysisScore.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace DirtBag.Modules {
 	class AnalysisScore {
-		public int Score { get; set; }
+		public double Score { get; set; }
 		public string Reason { get; set; }
 		public string ReportReason { get; set; }
 		public string ModuleName { get; set; }
@@ -15,7 +15,7 @@ namespace DirtBag.Modules {
 			Reason = "";
 			ModuleName = "";
 		}
-		public AnalysisScore(int score, string reason,string reportReason, string moduleName ) {
+		public AnalysisScore(double score, string reason,string reportReason, string moduleName ) {
 			Score = score;
 			Reason = reason;
 			ReportReason = reportReason;
